Add Tile/Internal.hs to enable a type for tiles
The List type for tiles is probably not going to be useful when it comes
to calculating paths and whether items move off the board or not.  Thus,
this patchset moves the List specific tile code into an Internal module
(to enable testing) and provides Tile as the API to tiles using the new
Tile type.